move environment variables at play level
All ansible modules are using two common environment variables to
determine if the deployment is containerized or not.
Setting those variables on each module call makes a lot of duplicate
code and also useless when the deployment isn't containeried.
Instead we should set the evironment variables at play level and only
on containerized deployment.

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>