rgw/multisite: add endpoints key to zones list
This removes the zone_endpoints_list fact since we can define it as the
zones fact to avoid duplicate facts and two calls to radosgw_zone module
(one with access/secret key and another one with the endpoints).
Now the endpoints key is merged into the zones definition so we can remove
the dedicated task which adds the endpoints to the zone.

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>