nfs: get org.ganesha.nfsd.conf from container
Since we need to revert 33bfb10, this is an alternative to initial approach.
We can avoid maintaining this file since it is present in container
image. The idea is to simply get it from the image container and write
it to the host.

Fixes: #6501

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>

Sensitive key data now hidden

Signed-off-by: Neelaksh Singh <neelaksh48@gmail.com>

Sensitive key data is now hidden

Sensitive key data now hidden