playbooks: use group name variables consistently
In many places, playbooks allow host group names such as 'mons' to be
overriden by defining a variable such as 'mon_group_name'. This is
very useful when ceph-ansible is sharing the ansible inventory with
other playbooks and other hosts which expect different group names.
But this was unfinished, with the variables used in some places but
hard-coded names used in others places, sometimes within the same
playbook.

This commit uses the group name variables everywhere so they can be
reliably used to overrride the host group names.

Signed-off-by: Stuart Grace <stuart.grace@bbc.co.uk>