Handle adoption when radosgw_address_block is comma delimited list
In cephadm-adopt.yml TASK "Update the placement of radosgw hosts"
passes module ceph_orch_apply embedded YAML via a block scalar.
This YAML creates a Ceph spec of service_type RGW. The networks
key of this spec supports either a list or a string. Without this
patch, the networks key of the spec will only contain a string.
With this patch a string is still set for the networks key, but
if Ansible var radosgw_address_block contains commas, then var
radosgw_address_block is split by those commas into a list and
the networks key of the spec will be set to a list.

Closes: https://issues.redhat.com/browse/RHCEPH-10418

Signed-off-by: John Fulton <fulton@redhat.com>