Handle radosgw hosts placement with non-default cluster name
In cephadm-adopt.yml TASK "Update the placement of radosgw hosts"
does not handle when Ansible var cluster is something other than
"ceph", unless this patch is used.

Update module ceph_orch_apply to support optional cluster
parameter using the same style as in module ceph_config.
The command is only extended to inclue the new keyring
and config options if cluster name is not ceph.

This patch is necessary to migrate older clusters which were
deployed when custom names were supported.

Closes: https://issues.redhat.com/browse/RHCEPH-10442

Signed-off-by: John Fulton <fulton@redhat.com>