Use grafana_server_addr to set prometheus networks list
When dashboard is enabled and module ceph_orch_apply is
called, if the grafana_server_addr is defined, then it
is used to populate the networks list in the spec of type
alertmanager. This is the case without this patch. With
this patch the same logic is applied to the spec of type
prometheus. Also, if the grafana_server_addr is a comma
delimited list, then a jinja2 expression handles passing
the IPs as a list.

Without this patch prometheus binds to all networks even
if grafana_server_addr is set which can create conflicts
with other services.

Fixes: https://bugzilla.redhat.com/2269009

Signed-off-by: John Fulton <fulton@redhat.com>
(cherry picked from commit 770a527a9ee6b34091798c6120235026ad1ecfa9)